requirements.txt
aiogram==3.5.0
streamlit==1.35.0
aiosqlite==0.20.0
python-dotenv==1.0.1